master Adding new configuration and functions
Was added two new functions:
- replace - help to find specific files and change his content
            This was necessary because the "sed" command was adding a new line on the end of the file
            And it was polluting the git's diff files.
            *This function will become a bin shell file in some moment.
- forever - sugar syntax to use an infinite looping with sleep time.

Changed GOPATH from $HOME/go to $HOME/gowork for standardizing the language folder of each programming language.

Added MYSQL's bin, it's "MAC" specific
The alias for the MYSQL Service are "MAC" specific

OBS: In the future probably will be interesting to have configuration files for a specific type of environment.
Or change everything to the "MAC" environment.